Memecah Class Employee
Memecahkan Kelas Employee menjadi 3 Kelas, Kelas utama Employee, dan 2 subclass, Employee Family, dan EmployeeSalary. Attribute dan method yang terlibat juga di pecah dari kelas employee
